refactor: update monitor repository and vital sign models

+++ lib/features/monitor/repository/monitor_repository.dart
@@ -26,7 +26,7 @@ class MonitorRepository {
     return mqtt.bpStream;
   }
 
-  Stream<int> getBPLiveStream() {
+  Stream<double> getBPLiveStream() {
     return mqtt.bpLiveStream;
   }
